pyproject.toml
[tool.poetry]
name = "front-page"
version = "0.1.0"
description = ""
authors = ["Alex Basso"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
